requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
dataclasses==0.6
discord.py==1.4.1
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.7.6
pbr==5.4.5
pycodestyle==2.6.0
pycparser==2.20
pylint==2.5.3
python-decouple==3.3
screeninfo==0.6.5
six==1.15.0
stevedore==3.2.0
toml==0.10.1
virtualenv==20.0.27
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
WTForms==2.3.1
yarl==1.5.1
